[#5]Add <funcname>.c files persing func
