+++ srcs/debug/debug_print.c
@@ -6,7 +6,7 @@
 /*   By: takira <takira@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/07 09:21:33 by takira            #+#    #+#             */
-/*   Updated: 2023/01/24 16:44:27 by takira           ###   ########.fr       */
+/*   Updated: 2023/01/24 16:44:30 by takira           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
