+++ srcs/tokenizer/operator_split.c
@@ -6,7 +6,7 @@
 /*   By: takira <takira@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/18 17:15:53 by takira            #+#    #+#             */
-/*   Updated: 2023/01/24 16:49:01 by takira           ###   ########.fr       */
+/*   Updated: 2023/01/24 17:03:12 by takira           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
